Move util under internal folder
