+++ cmd/va/collect.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/fatih/color"
 	"github.com/pkg/sftp"
-	"github.com/sailpoint-oss/sailpoint-cli/cmd/utilities"
+	"github.com/sailpoint-oss/sailpoint-cli/internal/util"
 	"github.com/spf13/cobra"
 	"github.com/vbauerster/mpb/v8"
 	"github.com/vbauerster/mpb/v8/decor"
@@ -55,7 +55,7 @@ func newCollectCmd() *cobra.Command {
 			log.SetOutput(p)
 
 			for credential := 0; credential < len(args); credential++ {
-				password, _ := utilities.PromptPassword(fmt.Sprintf("Enter Password for %v:", args[credential]))
+				password, _ := util.PromptPassword(fmt.Sprintf("Enter Password for %v:", args[credential]))
 				credentials = append(credentials, password)
 			}
