+++ cmd/va/troubleshoot.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 
 	"github.com/fatih/color"
-	"github.com/sailpoint-oss/sailpoint-cli/cmd/utilities"
+	"github.com/sailpoint-oss/sailpoint-cli/internal/util"
 	"github.com/spf13/cobra"
 )
 
@@ -27,7 +27,7 @@ func newTroubleshootCmd() *cobra.Command {
 			}
 			var credentials []string
 			for credential := 0; credential < len(args); credential++ {
-				password, _ := utilities.PromptPassword(fmt.Sprintf("Enter Password for %v:", args[credential]))
+				password, _ := util.PromptPassword(fmt.Sprintf("Enter Password for %v:", args[credential]))
 				credentials = append(credentials, password)
 			}
